requirements.txt
torch
pysimplegui
opencv-python
pillow